Use utc offset as courier channel
This fixes the issue where certain time zones will never
receive push notifications because the time zone names used by Rails are
different from those used by iOS (See thoughtbot/Tropos#185 for more
information)

This fixes it by not using the names, but the relative offset from UTC
as the channel name. Currently we broadcast to these channels when it's
9am PDT ["america-los_angeles", "america-tijuana", "america-phoenix"].

This PR changes it to broadcast to the channel minus-0700 instead. This
is a more reliable way of identifying time zones across platforms. And
it is the first step to allowing users to pick their preferred time.

thoughtbot/Tropos#187 is the PR to change this on the iOS side.